basic update is working
